+++ tools/env_loader.go
@@ -1,4 +1,4 @@
-package helpers
+package tools
 
 import (
 	"fmt"
